Add common tag suggestions
There are a few things to do, like fix the tests and remove the tags 
already included from consideration for most common one. The view test 
is broken because @repos is no longer just a Relation, but a structure 
with records and recommendations for them. The system tests are broken 
because the most common tag method gets nil sometimes.